DMP-715 Refactored notify() method
